+++ Restaurant/Restaurant/Classes/ApiClasses/AgentMembershipCash.cs
@@ -17,11 +17,11 @@ namespace Restaurant.Classes
     public class AgenttoPayCash
     {
 
-        public Nullable<int> agentMembershipCashId { get; set; }
-        public Nullable<int> subscriptionFeesId { get; set; }
-        public Nullable<int> cashTransId { get; set; }
-        public Nullable<int> membershipId { get; set; }
-        public Nullable<int> agentId { get; set; }
+        public Nullable<long> agentMembershipCashId { get; set; }
+        public Nullable<long> subscriptionFeesId { get; set; }
+        public Nullable<long> cashTransId { get; set; }
+        public Nullable<long> membershipId { get; set; }
+        public Nullable<long> agentId { get; set; }
         public Nullable<System.DateTime> startDate { get; set; }
         public Nullable<System.DateTime> endDate { get; set; }
         //public Nullable<int> updateUserId { get; set; }
@@ -29,7 +29,7 @@ namespace Restaurant.Classes
         public Nullable<System.DateTime> updateDate { get; set; }
         //public Nullable<int> createUserId { get; set; }
         public decimal Amount { get; set; }
-        public Nullable<int> pointId { get; set; }
+        public Nullable<long> pointId { get; set; }
         public string agentName { get; set; }
         public string agentcode { get; set; }
         public string agentcompany { get; set; }
@@ -49,7 +49,7 @@ namespace Restaurant.Classes
         public string agentpayType { get; set; }
         public bool agentcanReserve { get; set; }
         public string agentdisallowReason { get; set; }
-        public Nullable<int> agentresidentSecId { get; set; }
+        public Nullable<long> agentresidentSecId { get; set; }
         public string agentGPSAddress { get; set; }
 
         public string membershipName { get; set; }
@@ -75,18 +75,18 @@ namespace Restaurant.Classes
     }
     public class AgentMembershipCash
     {
-        public int agentMembershipsId { get; set; }
-        public Nullable<int> subscriptionFeesId { get; set; }
-        public Nullable<int> cashTransId { get; set; }
-        public Nullable<int> membershipId { get; set; }
-        public Nullable<int> agentId { get; set; }
+        public long agentMembershipsId { get; set; }
+        public Nullable<long> subscriptionFeesId { get; set; }
+        public Nullable<long> cashTransId { get; set; }
+        public Nullable<long> membershipId { get; set; }
+        public Nullable<long> agentId { get; set; }
         public Nullable<System.DateTime> startDate { get; set; }
         public Nullable<System.DateTime> EndDate { get; set; }
         public string notes { get; set; }
         public Nullable<System.DateTime> createDate { get; set; }
         public Nullable<System.DateTime> updateDate { get; set; }
-        public Nullable<int> createUserId { get; set; }
-        public Nullable<int> updateUserId { get; set; }
+        public Nullable<long> createUserId { get; set; }
+        public Nullable<long> updateUserId { get; set; }
         public byte isActive { get; set; }
         public decimal Amount { get; set; }
         public bool canDelete { get; set; }
@@ -101,12 +101,12 @@ namespace Restaurant.Classes
         public string transNum { get; set; }
         public Nullable<System.DateTime> payDate { get; set; }
         public byte membershipisActive { get; set; }
-        public int agentMembershipCashId { get; set; }
+        public long agentMembershipCashId { get; set; }
         public string subscriptionType { get; set; }
         public decimal discountValue { get; set; }
         public decimal total { get; set; }
         public string processType { get; set; }
-        public Nullable<int> cardId { get; set; }
+        public Nullable<long> cardId { get; set; }
         public string cardName { get; set; }
         public string docNum { get; set; }
         public string subscriptionTypeconv { get; set; }
@@ -126,7 +126,7 @@ namespace Restaurant.Classes
             return items;
         }
 
-        //public async Task<AgentMembershipCash> GetById(int itemId)
+        //public async Task<AgentMembershipCash> GetById(long itemId)
         //{
         //    AgentMembershipCash item = new AgentMembershipCash();
         //    Dictionary<string, string> parameters = new Dictionary<string, string>();
@@ -164,7 +164,7 @@ namespace Restaurant.Classes
             return await APIResult.post(method, parameters);
             //return  item.agentMembershipCashId
         }
-        //public async Task<int> delete(int itemId, int userId, Boolean final)
+        //public async Task<int> delete(long itemId, long userId, Boolean final)
         //{
         //    Dictionary<string, string> parameters = new Dictionary<string, string>();
         //    parameters.Add("itemId", itemId.ToString());
