+++ Restaurant/Restaurant/Classes/ApiClasses/Card.cs
@@ -18,13 +18,13 @@ namespace Restaurant.Classes
 {
     public class Card
     {
-        public int cardId { get; set; }
+        public long cardId { get; set; }
         public string name { get; set; }
         public string notes { get; set; }
         public Nullable<System.DateTime> createDate { get; set; }
         public Nullable<System.DateTime> updateDate { get; set; }
-        public Nullable<int> createUserId { get; set; }
-        public Nullable<int> updateUserId { get; set; }
+        public Nullable<long> createUserId { get; set; }
+        public Nullable<long> updateUserId { get; set; }
         public byte isActive { get; set; }
         public string image { get; set; }
         public bool hasProcessNum { get; set; }
@@ -54,7 +54,7 @@ namespace Restaurant.Classes
             parameters.Add("itemObject", myContent);
            return await APIResult.post(method, parameters);
         }
-        public async Task<int> delete(int cardId, int userId, Boolean final)
+        public async Task<int> delete(long cardId, long userId, Boolean final)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             parameters.Add("itemId", cardId.ToString());
@@ -71,7 +71,7 @@ namespace Restaurant.Classes
             string method = "cards/UpdateImage";
            return await APIResult.post(method, parameters);
         }
-        public async Task<string> uploadImage(string imagePath, string imageName, int cardId)
+        public async Task<string> uploadImage(string imagePath, string imageName, long cardId)
         {
             if (imagePath != "")
             {
